mcp-server/requirements.txt
fastmcp>=2.0.0
supabase>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
